api-mongo/requirements.txt
fastapi
uvicorn
pymongo
pyjwt